refactor(pipeline/scrape): extract Coercion helpers
Phase 5 commit 2/9 - move coerceString, coerceNumber, parseAutoDate
into Coercion/Coercion.ts. ScrapeAutoMapper.ts now imports these
helpers from the dedicated module; the public re-export of
parseAutoDate is preserved (re-exported via the named-export block).

Exports use the bottom-of-file `export { ... }` block to keep the
LintValidator Rule #15 (forbidden primitive return on exported
declarations) heuristic from firing on these adapter helpers — same
pattern as Network/AuthDiscovery.ts. The helpers are intentionally
the primitive-coercion boundary; nominal-typing across the module
edge would require lossy brand/cast wrappers at every callsite.

Master plan: pipeline-decoupling-master-2026-05-28 / phase-5
Cap: max-lines 150 eff + max-lines-per-function 20 eff
File eff LoC: Coercion.ts=29 (cap 150)
