+++ src/Scrapers/Pipeline/Mediator/Scrape/ScrapeAutoMapper.ts
@@ -4,13 +4,10 @@
  * field names automatically via BFS iterative search.
  */
 
-import moment from 'moment';
-
 import type { ITransaction } from '../../../../Transactions.js';
 import { TransactionStatuses, TransactionTypes } from '../../../../Transactions.js';
 import { ScraperErrorTypes } from '../../../Base/ErrorTypes.js';
 import {
-  KNOWN_DATE_FORMATS,
   PIPELINE_WELL_KNOWN_ACCOUNT_FIELDS as WK_ACCT,
   PIPELINE_WELL_KNOWN_API as WK_API,
   PIPELINE_WELL_KNOWN_BILLING as WK_BILLING,
@@ -33,6 +30,7 @@ import {
   type ScalarFieldHit,
   type UntypedValue,
 } from './AutoMapperFacade/AutoMapperTypes.js';
+import { coerceNumber, coerceString, parseAutoDate } from './Coercion/Coercion.js';
 
 export type { IMonthChunk } from '../Scrape/ScrapeReplayAction.js';
 export {
@@ -44,42 +42,6 @@ export {
 } from '../Scrape/ScrapeReplayAction.js';
 
 const LOG = getDebug(import.meta.url);
-
-/**
- * Coerce a field value to string, applying optional transform.
- * Numeric inputs are stringified so numeric YYYYMMDD dates survive.
- * @param val - Raw field value from findFieldValue.
- * @param transform - Optional string transform (e.g., parseAutoDate).
- * @param fallback - Fallback when val is missing.
- * @returns Coerced string.
- */
-function coerceString(
-  val: ScalarFieldHit,
-  transform?: (s: string) => string,
-  fallback = '',
-): string {
-  if (val === false) return fallback;
-  let s = '';
-  if (typeof val === 'string') s = val;
-  if (typeof val === 'number') s = String(val);
-  if (s === '') return fallback;
-  if (transform) return transform(s);
-  return s;
-}
-
-/**
- * Coerce a field value to number with fallback.
- * @param val - Raw field value from findFieldValue.
- * @param fallback - Fallback when val is not a number.
- * @returns Coerced number.
- */
-function coerceNumber(val: ScalarFieldHit, fallback: number): number {
-  if (typeof val === 'number') return val;
-  if (typeof val !== 'string') return fallback;
-  const parsed = Number(val);
-  if (Number.isNaN(parsed)) return fallback;
-  return parsed;
-}
 
 /**
  * Coerce a `findFieldValue` hit to a usable per-txn identifier.
@@ -538,17 +500,6 @@ function findFirstArray(obj: ApiRecord): readonly UntypedValue[] {
   });
   const hit = arrays.find((a): a is readonly UntypedValue[] => a !== false);
   return hit ?? [];
-}
-
-/**
- * Parse a date string using known bank formats.
- * @param dateStr - Raw date string from API response.
- * @returns ISO date string, or original if no match.
- */
-function parseAutoDate(dateStr: string): string {
-  const parsed = moment(dateStr, KNOWN_DATE_FORMATS, true);
-  if (parsed.isValid()) return parsed.toISOString();
-  return dateStr;
 }
 
 /** Shekel currency aliases from WK. */
